refactor: テーブル選択ロジックを useTableSelection フックに抽出
JdbcTableSelectorDialog と SqlTableInsertDialog で重複していた
toggle/toggleAll の Set<string> 操作を共通フックに統合。
